BREAKING(path): split up glob into multiple files
