+++ path/glob_to_regexp.ts
@@ -1,22 +1,19 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import { globToRegExp as _globToRegExp } from "./glob_to_regexp.ts";
-import { normalizeGlob as _normalizeGlob } from "./normalize_glob.ts";
-import { joinGlobs as _joinGlobs } from "./join_globs.ts";
-import { isGlob as _isGlob } from "../is_glob.ts";
+import type { GlobOptions } from "./_common/glob_to_reg_exp.ts";
+import { isWindows, OSType } from "./_os.ts";
 
-/**
- * @deprecated (will be removed in 0.215.0) Import from "std/path/posix/is_glob.ts"
- *
- * Test whether the given string is a glob
- */
-export const isGlob = _isGlob;
+import { globToRegExp as posixGlobToRegExp } from "./posix/glob_to_regexp.ts";
+import {
+  globToRegExp as windowsGlobToRegExp,
+} from "./windows/glob_to_regexp.ts";
 
-/**
- * @deprecated (will be removed in 0.215.0) Import from "std/path/posix/glob_to_regexp.ts"
- *
- * Convert a glob string to a regular expression.
+export type GlobToRegExpOptions = GlobOptions & {
+  os?: OSType;
+};
+
+/** Convert a glob string to a regular expression.
  *
  * Tries to match bash glob expansion as closely as possible.
  *
@@ -71,18 +68,11 @@ export const isGlob = _isGlob;
  *   fail to match `foobar.js`, even though `foobar` is not `foo`. Effectively,
  *   `!(foo|bar)` is treated like `!(@(foo|bar)*)`. This will work correctly if
  *   the group occurs not nested at the end of the segment. */
-export const globToRegExp = _globToRegExp;
-
-/**
- * @deprecated (will be removed in 0.215.0) Import from "std/path/posix/normalize_glob.ts"
- *
- * Like normalize(), but doesn't collapse "**\/.." when `globstar` is true.
- */
-export const normalizeGlob = _normalizeGlob;
-
-/**
- * @deprecated (will be removed in 0.215.0) Import from "std/path/posix/join_globs.ts"
- *
- * Like join(), but doesn't collapse "**\/.." when `globstar` is true.
- */
-export const joinGlobs = _joinGlobs;
+export function globToRegExp(
+  glob: string,
+  options: GlobToRegExpOptions = {},
+): RegExp {
+  return options.os === "windows" || (!options.os && isWindows)
+    ? windowsGlobToRegExp(glob, options)
+    : posixGlobToRegExp(glob, options);
+}
