+++ path/is_glob.ts
