+++ path/posix/is_glob.ts
@@ -1,0 +1,4 @@
+// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+// This module is browser compatible.
+
+export { isGlob } from "../is_glob.ts";
