+++ path/posix/mod.ts
@@ -41,4 +41,7 @@ export * from "./to_namespaced_path.ts";
 export * from "./common.ts";
 export * from "./separator.ts";
 export * from "../_interface.ts";
-export * from "../glob.ts";
+export * from "./glob_to_regexp.ts";
+export * from "./is_glob.ts";
+export * from "./join_globs.ts";
+export * from "./normalize_glob.ts";
